refactor: scss colors from constants
